chore: set validation if pizza exist

+++ src/services/pizzas/service.py
@@ -1,3 +1,5 @@
+from datetime import datetime
+
 from src.core.interfaces.services.pizzas.interface import IPizzaService
 from src.repositories.pizzas.repository import PizzasRepository
 
@@ -7,6 +9,18 @@ class PizzaService(IPizzaService):
     @classmethod
     async def create_pizza(cls, payload: dict, pizza_repo=PizzasRepository):
         data = payload.get("payload")
+        pizza_name = data.get("name")
+        query = {"name": pizza_name}
+        projection = {"_id": False}
+        pizza_data = await pizza_repo.find_one(query, projection)
+
+        if pizza_data:
+            return {
+                "status_code": 200,
+                "message": "Pizza already exist!. Try another pizza"
+            }
+
+        data.update({"created_at": datetime.now()})
         await pizza_repo.insert_one(data)
         return {
             "status_code": 201,
@@ -18,9 +32,9 @@ class PizzaService(IPizzaService):
         projection = {"_id": False}
         result = await pizza_repo.find_all({}, projection)
         return {
+            "result": result,
             "status_code": 200,
-            "message":"",
-            "result": result
+            "message": "",
         }
 
     @classmethod
@@ -28,20 +42,60 @@ class PizzaService(IPizzaService):
         projection = {"_id": False}
         pizza_name = payload.get("pizza_name")
         query = {"name": {"$regex": pizza_name, "$options": "i"}}
+
         result = await pizza_repo.find_one(query, projection)
-        return result
+        if not result:
+            return {
+                "result": [],
+                "status_code": 404,
+                "message": "Pizza not found",
+            }
+
+        return {
+            "result": result,
+            "status_code": 200,
+            "message": "",
+        }
 
     @classmethod
     async def update_pizza(cls, payload: dict, pizza_repo=PizzasRepository):
         pizza_name = payload.get("pizza_name")
         query = {"name": pizza_name}
+        projection = {"_id": False}
         new_data = payload.get("data")
+        new_data.update({"updated_at": datetime.now()})
+
+        result = await pizza_repo.find_one(query, projection)
+        if not result:
+            return {
+                "result": [],
+                "status_code": 404,
+                "message": "Pizza not found",
+            }
+
         await pizza_repo.update_one(query, new_data)
-        return "ok"
+        return {
+            "result": [],
+            "status_code": 200,
+            "message": "Pizza Created",
+        }
 
     @classmethod
     async def delete_pizza(cls, payload: dict, pizza_repo=PizzasRepository):
         pizza_name = payload.get("pizza_name")
         query = {"name": pizza_name}
+        projection = {"_id": False}
+
+        result = await pizza_repo.find_one(query, projection)
+        if not result:
+            return {
+                "result": [],
+                "status_code": 404,
+                "message": "Pizza not found",
+            }
         await pizza_repo.delete_one(query)
-        return "ok"
+        return {
+            "result": [],
+            "status_code": 200,
+            "message": "Pizza Deleted"
+        }
